+++ src/preprocess/debug.cpp
@@ -9,13 +9,13 @@
 
 namespace lvc {
 
-static LVC_FORCE_INLINE void dbgDrawCircle(cv::Mat& dst, cv::Point2f center, float diameter)
+static LVC_FORCE_INLINE void dbgDrawCircle(cv::Mat& dst, cv::Point2d center, double diameter)
 {
     cv::circle(dst, static_cast<cv::Point2i>(center), static_cast<int>(diameter / 2.0f), cv::Scalar(0, 255, 0), 1,
                cv::LineTypes::LINE_AA);
 }
 
-static LVC_FORCE_INLINE void dbgDrawSoildCircle(cv::Mat& dst, cv::Point2f center, float diameter)
+static LVC_FORCE_INLINE void dbgDrawSoildCircle(cv::Mat& dst, cv::Point2d center, double diameter)
 {
     cv::circle(dst, static_cast<cv::Point2i>(center), static_cast<int>(diameter / 2.0f), cv::Scalar::all(255.0),
                cv::LineTypes::FILLED, cv::LineTypes::LINE_AA);
@@ -40,7 +40,7 @@ void dbgDrawUsedArea(const Config& cfg, const cv::Mat& patch_size_map, int view_
 
     for (const auto& mi : mis) {
         auto idx = mi.getIndex();
-        float patch_size = static_cast<float>(patch_size_map.at<uchar>(idx.y, idx.x) + view_num);
+        double patch_size = static_cast<double>(patch_size_map.at<uchar>(idx.y, idx.x) + view_num);
         if (patch_size > cfg.getDiameter()) {
             patch_size = cfg.getDiameter();
         }
